refactor(website): split ImportVisual into a folder structure
Mirror the PipelineVisual layout: ImportVisual.tsx shell + barrel, with
the mapping row in components/, the ColumnMapping type in types/, and
MAPPINGS in data/.
