+++ packages/twenty-website/src/sections/product-feature/ImportVisual/index.ts
@@ -1,0 +1,1 @@
+export { ImportVisual } from './ImportVisual';
