+++ packages/twenty-website/src/sections/product-feature/ImportVisual/types/column-mapping.ts
@@ -1,0 +1,10 @@
+import { type ComponentType } from 'react';
+
+type FieldGlyph = ComponentType<{ size?: number; stroke?: number }>;
+
+export type ColumnMapping = {
+  Icon: FieldGlyph;
+  example: string;
+  field: string;
+  header: string;
+};
